Code of conduct for software:
An essential aspect of the code of conduct for our software is that it should primarily be used by individuals 14, and we hope the behavior will be followed. In addition, in no form will any threats, violence, racist comments, etc.
Plus, this software will be used by individuals who wish to learn and enrich their knowledge. 
- Roja K.