Code of Conduct:
An essential aspect of the code of conduct for our software is that it should primarily be used by individuals 14 and up. This is due to the fact that it will be easier for older teens to navigate and utilize the software. In this platform, users and collaborators will need to follow some behavioral rules such as: 
-no threats
-violence
-racist comments
- discriminatory behavior 
-etc
Plus,  my partner and I hope this software will be used by individuals who wish to learn and enrich their knowledge in addition to meeting new people, working collaboratively, and pursuing individual projects. 
- Roja K.